restructure to make sense
